Split out all of the class definitions in europi.py into their own files, create the new hardware namespace
