+++ software/firmware/hardware/pins.py
@@ -1,0 +1,15 @@
+# Pin assignments
+PIN_DIN = 22
+PIN_AIN = 26
+PIN_K1 = 27
+PIN_K2 = 28
+PIN_B1 = 4
+PIN_B2 = 5
+PIN_CV1 = 21
+PIN_CV2 = 20
+PIN_CV3 = 16
+PIN_CV4 = 17
+PIN_CV5 = 18
+PIN_CV6 = 19
+PIN_USB_CONNECTED = 24  # Does not work on Pico 2
+PIN_TEMPERATURE = 4
