+++ software/firmware/helpers.py
@@ -1,0 +1,3 @@
+def clamp(value, low, high):
+    """Returns a value that is no lower than 'low' and no higher than 'high'."""
+    return max(min(value, high), low)
